back-end/pyproject.toml
[tool.poetry]
name = "cs192-joggernaut-backend"
version = "0.3.0"
description = ""
authors = ["Enduro Team"]
license = "GNU General Public License"
readme = "../README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.7"
djangorestframework-types = "^0.9.0"
tzdata = "^2025.1"
pytz = "^2025.1"
django-oauth-toolkit = "^3.0.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.28.0"}
uvicorn = "^0.34.0"
ruff = "^0.11.2"
python-dotenv = "^1.0.1"
tblib = "^3.0.0"
coverage = "^7.7.1"
whitenoise = "^6.9.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # flake8-simplify
    "SIM",
    # flake8-return
    "RET",
    # flake8-bugbear
    "B",
    # isort
    "I",
    # pyupgrade
    "UP",
]
ignore = [
    # line too long
    "E501",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.coverage.paths]
source = ["."]

[tool.coverage.run]
command_line = "manage.py test --no-input"
omit = [
    "*/migrations/*",
    "*/django_backend/*",
    "*/manage.py",
    "*/urls.py",
    "*/__init__.py",
    "*/tests/test_*.py"
]

[tool.coverage.html]
directory = "api/tests/coverage/html"
